fix bug in adding info to errors
